@@ -1,6 +1,6 @@
 package com.acme.tictactoe.controller;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.acme.tictactoe.R;
